requirements.txt
# Base django 3.2 requirements
Django>=3.2.4,<4.0
django-debug-toolbar==3.2.4
gunicorn==20.1.0
psycopg2-binary==2.9.3
pytz==2021.1
sqlparse==0.4.2
whitenoise==6.0.0

# General
python-dotenv
requests

# Django LDAP 
django-auth-ldap
django-adminlte3

# Celery
celery
django_celery_results
django_celery_beat
redis
